Remove Inbox "deleting" and restoring
As discussed in #6

With the ability to empty and/or disable an inbox, "deleting" seems a
bit disingenuous as the inbox isn't completely removed (we need it to
make sure we don't generate that inbox for another user). There is also
the issue of restoring the inbox at a later date - it's pretty clunky.