Refactor search subqueries
The main part of this change is the way we do search subqueries.

Previously, we gave the search engine both subqueries and then called it
twice with a content_type look up. This would result in a redundand
subquery in both searches.

We now just do the search with each subquery directly.

A few minor changes too:

* Using model manager methods as our views do
* User ID is being used for cache key - saves a query in task

Touch #45