Protect against IDs and classes inside email bodies catching events
It might be possible for a malicious sender to choose IDs or classes
that are used by Inboxen's JS and influence how it is executed

Also, move the search JS into its own bundle. That uses a timer, so
rather than messing around with storing the timer ID somewhere we'll
only load that JS when searching.

Touch #82