Fix failing test attachment downloading tests
This fixes some test cases which were failing due to including the
method in the URL. This has been removed as of #210 however i neglected
to notice the failing tests. The URL no longer requires a method as
attachments can no longer be viewed, only downloaded for security reasons.

touch #210.